Move all build steps back into file containing defs
