+++ codegen/object.go
@@ -10,7 +10,6 @@ import (
 	"unicode"
 
 	"github.com/99designs/gqlgen/codegen/config"
-
 	"github.com/99designs/gqlgen/codegen/templates"
 	"github.com/pkg/errors"
 	"github.com/vektah/gqlparser/ast"
